move row logic to extra component
